Pull up window.innerHeight to flow through the whole app, and bugfixes
By setting the `window.innerHeight` as a `prop` in `_app.js`, this can flow down through props to any component that needs it. It's not perfect, but it allows all the math to work correctly. Also pulled up some of the height calculations to require fewer of them. And, fixed the heights of the `.grid--background` and `.content` elements, so they don't cause an additional scroll or poor footer placement on mobile