Add custom share image, because React can't handle image query parameters in the meta file
It's a known React bug, but the server-side rendering can't handle query
parameters in the meta head file; it turns `&`s into `&amp;`s, etc.

To bypass that, we've added a custom share image field to the content
model, and require that to be under 1MB so we don't have to handle it
here.

You *can* run an Express server using Next JS that will handle these
HTML entities and decode them, but Vercel (the host) won't allow that,
and also provides some custom optimization that other hosts wouldn't,
out-of-the-box. So, decided to not go that route.

Some open github issues, for reference:
- https://github.com/vercel/next.js/issues/2006
- https://github.com/facebook/react/issues/13838
- An example Express server fix: https://gist.github.com/stefl/1f8c246dd7ca9cb332ae41f68e80088d